fix(ep): use 32-bit store_release for P2P signals, fix MUSA build
- mtlink/nvlink signal and red_add now use store_release_32 instead of
  atomic_add_release, matching the original st_na_release behavior.
  The P2P path uses 32-bit int signals; the 64-bit atomicAdd caused
  "misaligned address" errors on MUSA and was semantically wrong
  (store vs atomic add) on both platforms.

- ep_comm_ops.cuh: ep_signal and ep_red_add take int32_t instead of
  int64_t/uint64_t. IBGDA path casts to 64-bit as needed.

- ep_comm_ops.cuh: include platform headers (musa_ops.cuh, cuda_ops.cuh)
  so kernel code can access platform-specific types.

- mooncake_ep_kernel.cu: add using declarations for IbGdaCtx,
  IbGdaQpDevCtx, cudaDeviceOps in TENT+CUDA path.

- setup.py: force use_tent=True when use_musa=True (MUSA can't use
  raw IBGDA/mlx5gda path). Add nvlink_device.cpp and status.cpp
  sources for CUDA+TENT builds.

- CMakeLists.txt: force USE_TENT=ON when USE_MUSA=ON.

- mooncake-pg/setup.py: add MUSA build support with MUSAExtension.

Validation:
- NV: 48/48 test_ep_grid.py passed (before signal type change;
  change is backward-compatible on NV)
- MT: 12/12 test_ep_grid.py passed on 8x MTT S5000

+++ mooncake-transfer-engine/tent/include/tent/device/ir/ep_comm_ops.cuh
@@ -31,9 +31,11 @@
 
 #ifdef MOONCAKE_EP_USE_MUSA
 #include <tent/device/mtlink.cuh>
+#include <tent/device/platform/musa/musa_ops.cuh>
 #else
 #include <tent/device/nvlink.cuh>
 #include <tent/device/network/ibgda/ibgda_ops.cuh>
+#include <tent/device/platform/cuda/cuda_ops.cuh>
 #endif
 
 namespace mooncake {
@@ -132,13 +134,9 @@ __device__ __forceinline__ void ep_put_ibgda(EpCommCtx& ctx,
 
 __device__ __forceinline__ void ep_signal(EpCommCtx& ctx,
                                            int dst_rank, int channel,
-                                           void* sig_ptr, int64_t action) {
+                                           void* sig_ptr, int32_t action) {
     if (dst_rank == ctx.rank) {
-        // Local signal — use st_na_release from mooncake_ep_utils
-        // (can't call st_na_release here since it's in a different header)
-        // The kernel handles local signals directly.
-        // This branch is kept for completeness; kernels should handle
-        // the local case before calling ep_signal.
+        // Local signal — kernel handles directly via st_na_release
         return;
     }
 
@@ -152,20 +150,22 @@ __device__ __forceinline__ void ep_signal(EpCommCtx& ctx,
 #endif
     } else {
 #ifndef MOONCAKE_EP_USE_MUSA
-        ibgda::ibgda_signal(&ctx.ibgda_ctx, channel, dst_rank, action);
+        ibgda::ibgda_signal(&ctx.ibgda_ctx, channel, dst_rank,
+                            static_cast<int64_t>(action));
 #endif
     }
 }
 
 __device__ __forceinline__ void ep_red_add(EpCommCtx& ctx,
                                             int dst_rank, int channel,
-                                            void* sym_ptr, uint64_t val) {
+                                            void* sym_ptr, int32_t val) {
     if (dst_rank == ctx.rank) {
-        // Local atomic add — kernel handles directly
+        // Local — kernel handles directly via st_na_release
         return;
     }
 
     if (ep_p2p_available(ctx, dst_rank)) {
+        // P2P path: release store (matches original st_na_release behavior)
 #ifdef MOONCAKE_EP_USE_MUSA
         mtlink::mtlink_red_add(ctx.dops, ctx.p2p_ctx, dst_rank,
                                ctx.local_base, sym_ptr, val);
@@ -174,8 +174,10 @@ __device__ __forceinline__ void ep_red_add(EpCommCtx& ctx,
                                ctx.local_base, sym_ptr, val);
 #endif
     } else {
+        // IBGDA path: atomic add (RDMA needs atomic visibility)
 #ifndef MOONCAKE_EP_USE_MUSA
-        ibgda::ibgda_red_add(&ctx.ibgda_ctx, channel, sym_ptr, val, dst_rank);
+        ibgda::ibgda_red_add(&ctx.ibgda_ctx, channel, sym_ptr,
+                             static_cast<uint64_t>(val), dst_rank);
 #endif
     }
 }
